Move veNft data to its own project type and change when it is fetched
