+++ src/components/veNft/VeNftDrawer.tsx
@@ -4,6 +4,13 @@ import { useMemo } from 'react'
 
 import VeNftContent from 'components/veNft/VeNftContent'
 
+import { VeNftContext, VeNftContextType } from 'contexts/v2/veNftContext'
+import { useVeNftBaseImagesHash } from 'hooks/veNft/VeNftBaseImagesHash'
+import { useVeNftLockDurationOptions } from 'hooks/veNft/VeNftLockDurationOptions'
+import { useVeNftResolverAddress } from 'hooks/veNft/VeNftResolverAddress'
+import { useVeNftUserTokens } from 'hooks/veNft/VeNftUserTokens'
+import { useVeNftVariants } from 'hooks/veNft/VeNftVariants'
+
 import { drawerStyle } from 'constants/styles/drawerStyle'
 
 export function VeNftDrawer({
@@ -13,10 +20,25 @@ export function VeNftDrawer({
   visible: boolean
   onClose: VoidFunction
 }) {
+  const veNftProjectName = 'veBanny'
+  const { data: lockDurationOptions } = useVeNftLockDurationOptions()
+  const { data: resolverAddress } = useVeNftResolverAddress()
+  const { data: variants } = useVeNftVariants()
+  const { data: userTokens } = useVeNftUserTokens()
+  const baseImagesHash = useVeNftBaseImagesHash()
   const memoizedDrawerStyle: Partial<DrawerProps> = useMemo(
     () => drawerStyle,
     [],
   )
+
+  const veNft: VeNftContextType = {
+    name: veNftProjectName,
+    lockDurationOptions,
+    resolverAddress,
+    variants,
+    userTokens,
+    baseImagesHash,
+  }
 
   return (
     <Drawer
@@ -25,7 +47,9 @@ export function VeNftDrawer({
       onClose={onClose}
       destroyOnClose
     >
-      <VeNftContent />
+      <VeNftContext.Provider value={veNft}>
+        <VeNftContent />
+      </VeNftContext.Provider>
     </Drawer>
   )
 }
